feat(rhc-templates): refactored set-up to support multiple pages and created form page
