--- apps/rhc-templates/src/app/globals.css
@@ -1,1 +1,0 @@
-/* global CSS here */
