+++ apps/rhc-templates/src/pages/form/index.tsx
@@ -1,0 +1,14 @@
+'use client';
+
+import { Logo } from '@rijkshuisstijl-community/components-react';
+
+export default function Form() {
+  return (
+    <>
+      <main>
+        <Logo organisation="Voorbeeld organisatie" subtitle="Voorbeeld sub-title"></Logo>
+        <div className="container page-content">Form</div>
+      </main>
+    </>
+  );
+}
