+++ apps/rhc-templates/src/pages/globals.css
@@ -1,4 +1,4 @@
-/* page CSS here */
+/* global CSS here */
 body {
   margin: 0;
   background-color: #f2f4f6;
