+++ apps/rhc-templates/src/pages/index.tsx
@@ -1,0 +1,7 @@
+export default function Empty() {
+  return (
+    <>
+      <p>Klik op de bovenstaande links om de verschillende templates te zien.</p>
+    </>
+  );
+}
