+++ proprietary/design-tokens/figma/figma.tokens.json
@@ -805,8 +805,6 @@
       }
     },
     "utrecht/document": {
-      "background-color": { "value": "{rhc.color.wit}" },
-      "color": { "value": "{rhc.color.zwart}" },
       "font-family": { "value": "{rhc.font-family.primary}" },
       "font-size": { "value": "{rhc.font-size.md}" },
       "line-height": { "value": "1.4" }
